make MultiHeadAttention agnostic to dtype (float32 vs. float16)
I was running with mixed precision and this layer caused a mismatch error (tf.float16 != tf.float32).

There might be a reason this needs to be float32 (for better precision). But if not, it would be nice to add the flexibility.

+++ tensorflow_addons/layers/multihead_attention.py
@@ -216,7 +216,7 @@ class MultiHeadAttention(tf.keras.layers.Layer):
 
         # Scale dot-product, doing the division to either query or key
         # instead of their product saves some computation
-        depth = tf.constant(self.head_size, dtype=tf.float32)
+        depth = tf.constant(self.head_size, dtype=query.dtype)
         query /= tf.sqrt(depth)
 
         # Calculate dot product attention
